moved demo() to lib_disparity.py
