+++ stereoigev/lib_disparity.py
@@ -4,6 +4,7 @@ Wrapper library introduced in forked version.
 """
 
 import argparse
+import glob
 import os
 from dataclasses import dataclass, field
 from pathlib import Path
@@ -12,6 +13,7 @@ import cv2
 import numpy as np
 import torch
 from PIL import Image
+from tqdm import tqdm
 
 from stereoigev.igev_stereo import IGEVStereo
 from stereoigev.utils import InputPadder
@@ -80,3 +82,36 @@ class DisparityCalculator:
         torch_image1 = as_torch_img(bgr1, is_BGR_order=True)
         torch_image2 = as_torch_img(bgr2, is_BGR_order=True)
         return self.calc_by_torch_image(torch_image1, torch_image2)
+
+
+def demo(args: argparse.Namespace):
+    """
+    save disparity files using left_imgs, right_imgs
+
+    args: in Namespace format
+        see details in command line help(-h).
+    """
+
+    disparity_calculator = DisparityCalculator(args=args)
+    output_directory = Path(args.output_directory)
+    output_directory.mkdir(exist_ok=True)
+
+    with torch.no_grad():
+        left_images = sorted(glob.glob(args.left_imgs, recursive=True))
+        right_images = sorted(glob.glob(args.right_imgs, recursive=True))
+        print(f"Found {len(left_images)} images. Saving files to {output_directory}/")
+
+        for imfile1, imfile2 in tqdm(list(zip(left_images, right_images))):
+            disparity = disparity_calculator.calc_by_name(imfile1, imfile2)
+            file_stem = Path(imfile1).stem
+            filename = output_directory / f"{file_stem}.png"
+
+            if args.save_numpy:
+                np.save(output_directory / f"{file_stem}.npy", disparity)
+            disp = np.round(disparity * 256).astype(np.uint16)
+            cv2.imwrite(
+                str(filename),
+                cv2.applyColorMap(cv2.convertScaleAbs(disp, alpha=0.01), cv2.COLORMAP_JET),
+                [int(cv2.IMWRITE_PNG_COMPRESSION), 0],
+            )
+            print(f"saved {filename}")
